Chore: remove dependency on dill (#103)
Early in pysaliency, we used dill to save datasets and fixations
to pickle files. Since then, hdf5 turned out to be more useful
for this and I want to get rid of the dill feature. For now
it's only deprecated and dill is not required anymore to run
pysaliency unless you want to load pickle files. But the
feature is still tested. In a later version, we can get rid of it
for good.

Signed-off-by: Matthias Kümmerer <matthias@matthias-k.org>